give verificator methods its own namespace
