--- app/controllers/decidim/action_delegator/admin/application_controller.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ActionDelegator
-    module Admin
-      # This controller is the abstract class from which all other controllers of
-      # this engine inherit.
-      #
-      class ApplicationController < Decidim::Admin::ApplicationController
-      end
-    end
-  end
-end
