+++ app/controllers/decidim/action_delegator/verification/admin/delegations_controller.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ActionDelegator
+    module Verification
+      module Admin
+        class DelegationsController < ActionDelegator::Verification::Admin::ApplicationController
+          include NeedsPermission
+
+          layout "decidim/admin/users"
+
+          def index
+            # some permissions
+          end
+        end
+      end
+    end
+  end
+end
