+++ app/views/decidim/action_delegator/verification/admin/delegations/index.html.erb
