+++ config/locales/en.yml
@@ -3,6 +3,14 @@ en:
   decidim:
     action_delegator:
       name: Decidim ActionDelegator cooperatives module
+    authorization_handlers:
+      admin:
+        delegations_verifier:
+          help:
+          - A verification method for delegate voting
+      delegations_verifier:
+        explanation: Verified users with this method can delegate votes to other users
+        name: Delegations Verifier
   layouts:
     decidim:
       user_profile:
