--- lib/decidim/action_delegator/admin.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ActionDelegator
-    # This module contains all the domain logic associated to Decidim's ActionDelegator
-    # component admin panel.
-    module Admin
-    end
-  end
-end
