+++ lib/decidim/action_delegator/verification/admin.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ActionDelegator
+    module Verification
+      # This module contains all the domain logic associated to the
+      # delegations verfier authorization admin panel.
+      module Admin
+      end
+    end
+  end
+end
