+++ lib/decidim/action_delegator/workflow.rb
@@ -3,6 +3,6 @@
 require "decidim/verifications"
 
 Decidim::Verifications.register_workflow(:delegations_verifier) do |workflow|
-  # workflow.engine = Decidim::ActionDelegator::VerificationEngine
-  workflow.admin_engine = Decidim::ActionDelegator::AdminEngine
+  workflow.engine = Decidim::ActionDelegator::Verification::Engine
+  workflow.admin_engine = Decidim::ActionDelegator::Verification::AdminEngine
 end
